BCT_Backend/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.0.9
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
greenlet==3.1.0
idna==3.10
mssql-django==1.5
numpy==2.1.1
pandas==2.2.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.0.8
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.34
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3